refactor: fold docs_repo into github_utils with public API at the top
